CrowdPulse/backend/requirements.txt
flask
flask-cors
praw
vaderSentiment
google-api-python-client 
python-dotenv
textblob
